Testing forecast and moving code around
