+++ test/templates/forecast-fail.yml
@@ -37,6 +37,24 @@ Resources:
               - !Sub "arn:${AWS::Partition}:s3:::${RetainBucket}"
               - !Sub "arn:${AWS::Partition}:s3:::${RetainBucket}/*" 
 
+  Instance1:
+    Type: AWS::EC2::Instance
+    Properties:
+      InstanceType: t3.micro 
+      SecurityGroups: [!Ref 'Instance1SecurityGroup']
+      KeyName: abc
+      ImageId: "does.not.exist"
+
+  Instance1SecurityGroup:
+    Type: AWS::EC2::SecurityGroup
+    Properties:
+      GroupDescription: Enable SSH access via port 22
+      SecurityGroupIngress:
+      - IpProtocol: tcp
+        FromPort: 22
+        ToPort: 22
+        CidrIp: 54.240.196.187/1
+
 
 Outputs:
